Auditing of Setting...View...Invoice.
Default Public Template (currently Invoice_Web) - 'his is the HTML template that the client will see online prior to payment. The template has a pay-now button. The client must log in having been assigned observer role status in order to see this html invoice template. Different HTML Templates can be created in this folder and chosen in this dropdown.

It is used in: 'views/invoice/template/invoice/public/Invoice_Web.php (subsequent to client gateway selection from inv/view) and also InvController/url_key function that receives the url_key and gateway query parameters in the Url from inv/view. This HTML template holds the pay-now button with the chosen gateway (passed from inv/view) which at this point cannot be changed. If the payment is successful the template and therefore the pay-now button will reflect as paid.'

Mark Invoices Sent Copy  - Clients do not have access to draft invoices. Mark a copied invoice as sent so that the client can view it. Normally used for testing purposes. By default copied invoices are marked as draft and therefore cannot be viewed by the client online.