Post update testing 4
A client with invView permission can edit the sales order from their side with their purchase order:

1) item identification number
2) line number

so that complete matching at source occurs.

PeppolHelper exceptions are raised as close to the source of the error as possible and preferably not in the src\Invoice\Ubl validation sections of each UBL class.

src\Invoice\Ubl\InvoiceLine is currently not being used. The invoiceline as mentioned previously is being generated by the PeppolHelper build_invoice_lines_array function.