Installation Testing 1:  active_user account
Avoid using:

$this->user_service->getUser() with a @psalm_suppress statement preceding it.

Make sure that the user_id that an invoice is linked to, is retrieved from a client with a user account by using the

active_user function.

This ensures that payments of invoices can be chased to the user account ie. the individual logged in.

An accountant making payments for invoices that we have sent to him will have clients assigned to his user account and he will be able to make payments for these clients under his user account when logged in.

The assumption is that all clients that are created on the system can only have one and not two user accounts setup hence the use of the userclient table to track these assignments.