Installation Testing 1: Inv Index Access
Only allow the inv index to be accessed if there are clients with active user accounts ie. the clients that have been created have been linked to an active user account. ie. signed up and linked.

Next: Change the Product Form to a Yiisoft based Form with dependency on the required parameter for each field to reveal errors ... using the FormHydrator. Currently the forms are dependent on the required html parameter.