Integration Testing 1: Realign with the demo
Using the structure similar to the demo, I have moved the project into a folder called invoice.  This aligns it with the demo so that the blog and blog-api can be copied into the root folder and any adjustments that have been made to the demo can be simply copied into the root folder structure.

This makes studying the changes that are made to the demo easier by running them alongside the invoice folder.

Two Separate virtual localhosts can then be made for the blog and blog-api. If the config/common/param.php file in each of the blog and blog-api is adjusted to match the invoice database cycle setup in its config/common/param.php then both can use the same database instead of the sqlite database.

Next: Iimplementing annotations for the routes. So changes will be made to how currentRoute is implemented. This is to keep up with the demo.